refactor: Entity 클라스 변경 및 맵핑 추가
