--- BE/airdnb/src/main/java/com/team01/airdnb/accommadation/Image.java
@@ -1,18 +1,0 @@
-package com.team01.airdnb.accommadation;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-@Entity
-public class Image {
-  @Id
-  @GeneratedValue
-  private Long id;
-  @ManyToOne
-  @JoinColumn(name = "accommodation_id")
-  private Accommodation accommodation;
-  private String imagePath;
-}
