--- BE/airdnb/src/main/java/com/team01/airdnb/accommadation/Wishlist.java
@@ -1,21 +1,0 @@
-package com.team01.airdnb.accommadation;
-
-import com.team01.airdnb.user.User;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-@Entity
-public class Wishlist {
-  @Id
-  @GeneratedValue
-  private Long id;
-  @ManyToOne
-  @JoinColumn(name = "user_id")
-  private User user;
-  @ManyToOne
-  @JoinColumn(name = "accommodation_id")
-  private Accommodation accommodation;
-}
