+++ BE/airdnb/src/main/java/com/team01/airdnb/comment/Comment.java
@@ -1,0 +1,35 @@
+package com.team01.airdnb.comment;
+
+import com.team01.airdnb.accommadation.Accommodation;
+import com.team01.airdnb.user.User;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "comments")
+public class Comment {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+  private Double score;
+  private String content;
+  private LocalDateTime createdAt;
+
+  @ManyToOne
+  @JoinColumn(name = "user_id")
+  private User user;
+
+  @ManyToOne
+  @JoinColumn(name = "accommodation_id")
+  private Accommodation accommodation;
+}
