+++ BE/airdnb/src/main/java/com/team01/airdnb/image/Image.java
@@ -1,26 +1,27 @@
-package com.team01.airdnb.accommadation;
+package com.team01.airdnb.image;
 
+import com.team01.airdnb.accommadation.Accommodation;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
 @Getter
 @Setter
-public class Comment {
+@Table(name = "images")
+public class Image {
   @Id
-  @GeneratedValue
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
+  private String imagePath;
+
   @ManyToOne
   @JoinColumn(name = "accommodation_id")
   private Accommodation accommodation;
-  private String userId;
-  private Double score;
-  private String content;
-  private LocalDateTime createdAt;
 }
